[C++] Warnings cleanup.

+++ sbe-samples/src/main/cpp/OtfExample.cpp
@@ -449,10 +449,11 @@ int main(int argc, char **argv)
     std::uint64_t actingVersion = headerDecoder.getSchemaVersion(buffer);
     std::uint64_t blockLength = headerDecoder.getBlockLength(buffer);
 
-    std::shared_ptr<std::vector<Token>> messageTokens = irDecoder.message(templateId, actingVersion);
+    std::shared_ptr<std::vector<Token>> messageTokens = irDecoder.message(
+        static_cast<int>(templateId), static_cast<int>(actingVersion));
 
-    const std::size_t result =
-        OtfMessageDecoder::decode(messageBuffer, length, actingVersion, blockLength, messageTokens, tokenListener);
+    const std::size_t result = OtfMessageDecoder::decode(
+        messageBuffer, length, actingVersion, blockLength, messageTokens, tokenListener);
 
     std::cout << "result = " << result << "\n";
